s3_bucket_subscription: constrain policy to respect filter
`filter_prefix` defines the subset of objects in an S3 bucket which
should invoke our lambda. There is no reason for our lambda to have read
access to objects which do not match this filter, so we can append it to
the corresponding policy to ensure we restrict our permissions to the
appropriate prefix set.