cloudwatch_logs_subscription: introduce `allow_all_log_groups` (#4)
Calling the module with a large number of log groups will raise an error
due to the lambda policy overrunning AWS limits. This commit introduces
an option which replaces the default behavior of creating one permission
per log group, with a single permission allowing all log groups in an
account to invoke our lambda function.