lambda: add lambda_s3_custom_rules variable (#17)
This commit adds an option to configure the `S3_CUSTOM_RULES`
environment variable used by the Observe lambda function.

This environment variable is a base64-encoded JSON array of "rules".
Each rule is composed of a regular expression and a map of HTTP headers.
The S3 URI of a created object will be sequentially compared against the
provided regular expressions. On the first match, we will apply the
associated HTTP headers to the outbound request towards Observe. If no
match is found, we fall back to inferred rules.