snapshot: allow extending as well as overriding actions
We have a default set of actions which are supported by our lambda. This
list keeps getting extended over time as we pile on more functionality,
but we can't always guarantee that it'll be representative of a lambda's
capabilities.

For example, if we're running `terraform-aws-lambda` v0.6.0, with a
lambda version of `v1.0.20210923`, we may know that the binary supports
`autoscaling:Describe*`, even if the terraform module does not specify
it explicitly.

In this case, we want to extend the action list, rather than override
it. This commit adds a variable `include`, which is appended to
`action`. This allows the following possibilities and combinations
thereof:

- defining `action` will override the base set of actions
- defining `include` will extend the existing set of actions
- defining `exclude` will filter out a subset of actions