feat!: declare variables as nullable (#46)
This commit introduces the `nullable` argument to variable blocks. As a
result of this change, the minimum provider version is bumped to 1.1.0.

The `nullable` argument allows callers of this module to override
variables while still being able to fall back to our defaults, e.g.

```
    variable "memory_size" {
      type    = number
      default = null
    }

    module "observe_lambda" {
      source  = "observeinc/lambda/aws"
      ...
      memory_size      = var.memory_size
    }
```

Previously, passing `null` to `memory_size` in our module would error
out, since the variable was expected to be a string. Now, passing
`null` will cause terraform to fall back to the underlying default.