feat(s3_bucket_subscription)!: revert `bucket_arns` to count (#51)
This reverts commit a13b8fd75c8f1e882da0d2869a39acd06930a9e0.

A drawback of using `set` and `for_each` is that the keys must be known
at apply-time in order to be used to address resource instances. This is
a fundamental limitation:
https://developer.hashicorp.com/terraform/language/meta-arguments/for_each#limitations-on-values-used-in-for_each

In the context of this module, users pass in a collection a buckets.
There is no obvious key that can be used to satisfy the current
constraint for `for_each`. In the absence of an obvious candidate, users
can enumerate a list and provide an index, at which point we're back to
the same limitations that `count` has. At that point, it is better to
revert and maintain backward compatibility with 1.x, and defer changes
that attempt to retain stable resource identifiers for future major
revisions.

This snafu should have been caught during testing - unfortunately we do
not run plan or apply as part of our current CI jobs, and manual testing
focused on the upgrade path rather than clean apply.