fix(s3_bucket_subscription): avoid race on subscription (#53)
Currently on applying this module, users may hit the following error:

```
│ Error: putting S3 Bucket Notification Configuration: InvalidArgument: Unable to validate the following destination configurations
│       status code: 400, request id: 9DXT97CA77A3MTPX, host id: 9R+2v/kbrAp57YCPbITdTv4XUzLSq7rlzANk3rIQjRSNsKJxaA4wsPIgNNr8UmaAKNOaEPPkS0wO/A1mEOS4xw==
```

This appears to be due to a race between applying
`aws_lambda_permission` and `aws_s3_bucket_notification`. This commit
ensures the S3 bucket has permission to invoke the lambda prior to
assigning the lambda as a target for bucket notifications.