feat: add s3_bucket module
When configuring a bucket to trigger the Observe Lambda, there are many
options that are commonly useful. This includes setting lifecycle rules
or configuring permissions for services such as CloudWatch or CloudTrail
to write to a prefix within the bucket.

Ths commit introduces the `s3_bucket` module, which is modelled after
our current s3 bucket configuration in the `observeinc/collection/aws`
module.