feat: support pulling from cloudwatch metrics (#58)
This commit introduces a new submodule, `cloudwatch_metrics`. Similar to
our API snapshot functionality, we set up a eventbridge rule that
periodically triggers the lambda with a custom payload.

This payload tells the lambda what metrics to scrape for. Callers of
this module can configure metric collection through the `period`,
`interval`, `delay` and `filters` variables.

Users must provide a list of filters. Each filter must have a namespace,
and an optional set of metric names and dimensions.

Co-authored-by: Ben Drucker <bvdrucker@gmail.com>