fix: address snapshot race
We need to ensure we only invoke the snapshot lambda once we have the
appropriate permission. Since we depend on the eventbridge target, we
should ensure the target is only installed after the lambda permission
has been configured.

We also update the version constraint on `aws` to reflect the use of
`aws_lambda_invocation` which was only introduced in 3.72.0